fix(state): keep additive registry internal
