+++ src/state/openclaw-state-db-contract.ts
@@ -23,30 +23,6 @@ export const LAZY_ADDITIVE_STATE_TABLES = [
   "worker_environment_ssh_fallback_ports",
 ] as const;
 export const LAZY_ADDITIVE_STATE_INDEXES = [...FIRST_USE_STATE_INDEXES] as const;
-type BareNullableSqliteDatatype = "ANY" | "BLOB" | "INT" | "INTEGER" | "REAL" | "TEXT";
-type LazyAdditiveStateColumnDefinition = {
-  columnName: string;
-  dataType: BareNullableSqliteDatatype;
-  tableName: string;
-};
-// Added after v6 shipped. Every definition stays bare and nullable so older v6
-// writers can omit it safely when a newer build has already ensured the column.
-export const CLAW_LAZY_ADDITIVE_STATE_COLUMN_DEFINITIONS = [
-  { columnName: "bootstrap_content_digest", dataType: "TEXT", tableName: "claw_installs" },
-  { columnName: "bootstrap_source_path", dataType: "TEXT", tableName: "claw_installs" },
-  { columnName: "extension_adapter_identity", dataType: "TEXT", tableName: "claw_package_refs" },
-  { columnName: "extension_detected_format", dataType: "TEXT", tableName: "claw_package_refs" },
-  { columnName: "extension_format", dataType: "TEXT", tableName: "claw_package_refs" },
-  { columnName: "extension_id", dataType: "TEXT", tableName: "claw_package_refs" },
-  { columnName: "extension_mapped_json", dataType: "TEXT", tableName: "claw_package_refs" },
-  { columnName: "extension_unavailable_json", dataType: "TEXT", tableName: "claw_package_refs" },
-  { columnName: "shared_host", dataType: "INTEGER", tableName: "worker_environments" },
-  { columnName: "run_end_cleanup_json", dataType: "TEXT", tableName: "worktrees" },
-] as const satisfies readonly LazyAdditiveStateColumnDefinition[];
-export const CLAW_LAZY_ADDITIVE_STATE_COLUMNS: readonly string[] =
-  CLAW_LAZY_ADDITIVE_STATE_COLUMN_DEFINITIONS.map(
-    ({ columnName, tableName }) => `${tableName}.${columnName}`,
-  );
 /** Maximum time one synchronous SQLite call may wait for a lock. */
 export const OPENCLAW_SQLITE_BUSY_TIMEOUT_MS = 5_000;
 /** User-facing guide for schema refusals; lives here so error sites avoid import cycles. */
