+++ src/state/openclaw-state-db-maintenance.ts
@@ -10,7 +10,6 @@ import {
   readSqliteUserVersion,
 } from "../infra/sqlite-user-version.js";
 import {
-  CLAW_LAZY_ADDITIVE_STATE_COLUMNS,
   OPENCLAW_DATABASE_SCHEMA_DOCS_URL,
   LAZY_ADDITIVE_STATE_TABLES,
   OPENCLAW_STATE_SCHEMA_VERSION,
@@ -19,7 +18,23 @@ import {
 import { resolveOpenClawStateSqlitePath } from "./openclaw-state-db.paths.js";
 import { OPENCLAW_STATE_SCHEMA_SQL } from "./openclaw-state-schema.js";
 
-export { CLAW_LAZY_ADDITIVE_STATE_COLUMNS } from "./openclaw-state-db-contract.js";
+/**
+ * Additive Claw provenance columns that only a writable open can ensure. A
+ * same-version database written before them stays readable so read-only
+ * planning surfaces are not refused before they can report anything.
+ */
+export const CLAW_LAZY_ADDITIVE_STATE_COLUMNS = [
+  "claw_installs.bootstrap_content_digest",
+  "claw_installs.bootstrap_source_path",
+  "claw_package_refs.extension_adapter_identity",
+  "claw_package_refs.extension_detected_format",
+  "claw_package_refs.extension_format",
+  "claw_package_refs.extension_id",
+  "claw_package_refs.extension_mapped_json",
+  "claw_package_refs.extension_unavailable_json",
+  "worker_environments.shared_host",
+  "worktrees.run_end_cleanup_json",
+] as const;
 
 const OPENCLAW_STATE_MAINTENANCE_SCHEMA_COMPATIBILITY = {
   allowCompatibleAdditiveColumns: true,
